requirements.txt
requests
python-dotenv
pandas
openai

